Revert "Update sound (#31)"
This reverts commit 300c97a8732367d0dbb1f04d3c74fdac7b2f515b.
