--- Resources/GameFoundationSoundGroup.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ba341ef36ac52354280089d6b9325b69
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
