--- Scripts/Utilities/SoundServices/IAudioManager.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: b6dc4901bb124ddd81cde70c187f3dad
-timeCreated: 1677742738
