--- Scripts/Utilities/SoundServices/MasterAAASoundWrapper.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 532fbbdec6a94f6b9b51ff857a3eb64c
-timeCreated: 1677141046
