--- Scripts/Utilities/SoundServices/SoundManagerInstaller.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 277131a28ab1482698b763b2f460786b
-timeCreated: 1677752531
