--- Scripts/Utilities/SoundServices/SoundModel.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: b1f0498004d54da7a56423aec5247812
-timeCreated: 1677745895
